Fix handling of tokens following a double dash.
This commit prevents the parser from processing any tokens that follow a
double dash.  Previously, the parser would try to match the tokens
against acceptable arguments and options to the current command
following the double dash.  This may result in modified tokens being
returned from the lex operation when the expected behavior is to leave
the tokens unmodified.

The fix is for the lex operation to keep track of whether or not a
double dash was encountered and, if so, return "unparsed" tokens for the
remaining tokens.

This fixes [CLI #7863](https://github.com/dotnet/cli/issues/7863).